[5000 Monkecoin] IPC and RDS do not work together.
As it states in the title, the species Integrated Positronic Chassis and the quirk Reality Dissociation Syndrome do not work together, as a byproduct of IPCs having general hallucination immunity only being affected whenever a global hallucination is triggered, no other source appears to induce such.

I'm not sure how much to give in regards to something seemingly simple, but I'd rather be generous, so I settled myself on 5000 monkecoin. I would also need to know how to perform the exchanging of monkecoins when this is fixed

i just made a immunity bypasser for mobs